PF Fix theming error
